fusetest: move test into respective files
